+++ mount/fmountat_linux.go
@@ -14,7 +14,7 @@
    limitations under the License.
 */
 
-package sys
+package mount
 
 import (
 	"runtime"
